cleaned up the code, reorganised the src folder, should be ready for react routing
