+++ IzvorniKod/frontend/package.json
@@ -14,6 +14,7 @@
     "frontend": "file:",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.28.0",
     "react-scripts": "5.0.1",
     "ts-node": "^10.9.2",
     "web-vitals": "^2.1.4"
