+++ IzvorniKod/frontend/src/App.test.tsx
@@ -1,19 +1,9 @@
 import React from 'react';
-import './App.css';
-import Heading from './components/heading';
-import Footer from './components/footer';
-import LoginSignUp from './components/LoginSignUp';
+import { render, screen } from '@testing-library/react';
+import App from './App';
 
-function App() {
-  return (
-    <div className="App">
-      <Heading />
-      <div className="content">
-        <LoginSignUp />
-      </div>
-      <Footer />
-    </div>
-  );
-}
-
-export default App;
+test('renders learn react link', () => {
+  render(<App />);
+  const linkElement = screen.getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
+});
