+++ IzvorniKod/frontend/src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import Heading from './components/heading';
 import Footer from './components/footer';
-import NewsFeed from './components/NewsFeed';
+import NewsFeed from './pages/NewsFeed';
 
 function App() {
   return (
