+++ IzvorniKod/frontend/src/assets/R.png
