+++ IzvorniKod/frontend/src/components/newsfeed.css
