+++ IzvorniKod/frontend/src/pages/NewsFeed.tsx
